[fix] Fix the cppcheck error '(error) Shifting signed 32-bit value by 31 bits is undefined behaviour'.

+++ stm32f4xx_usb_fs.c
@@ -1465,7 +1465,7 @@ void OTG_FS_IRQHandler(uint8_t irq __UNUSED, // IRQ number
     val &= intmsk;
 
 	for (i = 0; i < 32; i++) {
-		if (!(intsts & (1 << i)) || !(intmsk & (1 << i)))
+		if (!(intsts & ((uint32_t)1 << i)) || !(intmsk & ((uint32_t)1 << i)))
 			continue;
         usb_fs_isr_handlers_count[i]++;
 		if (usb_fs_isr_handlers[i]){
